backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.8.4
Pillow==10.4.0
python-dotenv==1.0.0
psycopg[binary]==3.2.3
pyjwt==2.8.0
bcrypt==4.1.2